+++ chat-ui/src/app/api/user/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { getAccessToken } from '@/app/utils/util';
+import { getAccessToken } from '@/app/utils/utils';
 
 export async function GET(req: Request) {
   const accessToken = await getAccessToken(req);
@@ -13,17 +13,11 @@ export async function GET(req: Request) {
     });
 
     if (!res.ok) {
-      return NextResponse.json(
-        { error: 'Failed to fetch user info' },
-        { status: res.status }
-      );
+      return NextResponse.json({ error: 'Failed to fetch user info' }, { status: res.status });
     }
 
     return NextResponse.json(await res.json());
   } catch (err) {
-    return NextResponse.json(
-      { error: 'Unexpected error', details: String(err) },
-      { status: 500 }
-    );
+    return NextResponse.json({ error: 'Unexpected error', details: String(err) }, { status: 500 });
   }
 }
