+++ chat-ui/src/app/components/SearchResult.tsx
@@ -1,3 +1,5 @@
+import { Result } from '@/app/models/result';
+
 export default function SearchResult({
   results,
 }: Readonly<{
@@ -9,7 +11,7 @@ export default function SearchResult({
 
   return (
     <div className="space-y-4">
-      {results.map((r, i) => (
+      {results.map((r) => (
         <div key={r.id} className="p-3 rounded bg-gray-50">
           <p className="font-medium text-blue-600">You: {r.query}</p>
           <p className="mt-1">Answer: {r.answer}</p>
