+++ chat-ui/src/app/models/history.ts
@@ -1,4 +1,4 @@
-interface HistoryItem {
+export interface HistoryItem {
   id: number;
   title: string;
 }
