+++ chat-ui/src/app/models/result.ts
@@ -1,4 +1,4 @@
-interface Result {
+export interface Result {
   id: number;
   query: string;
   answer: string;
