+++ chat-ui/src/app/models/token.ts
@@ -1,0 +1,5 @@
+export interface TokenInfo {
+  access_token: string;
+  refresh_token: string;
+  expires_in?: number;
+}
